Refactor to break into modules.
