+++ automator/backend/requirements.txt
@@ -1,3 +1,3 @@
+fastapi[standard]
 google-cloud-resource-manager
 google-cloud-asset
-fastapi[standard]
