+++ automator/backend/services/resources_manager.py
